# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Maestro Symphony",
    "description": "Maestro Symphony is a fast, mempool-aware Bitcoin indexer and API server. It supports mainnet and testnet4, providing data for UTXOs, metaprotocols and much more. Designed for developers who need reliable, mempool-aware blockchain data.",
    "contact": {
      "name": "jharper@gomaestro.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.txt"
    },
    "version": "v0.0.1"
  },
  "paths": {
...
    "/addresses/{address}/runes/utxos/{rune}": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Rune UTxOs by Address and Rune",
        "description": "Returns UTxOs controlled by the provided address which contain the provided rune, sorted by height.",
        "operationId": "addresses_specific_rune_utxos",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Bitcoin address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "tb1qphcdyah2e4vtpxn56hsz3p6kapg90pl4x525kc"
          },
          {
            "name": "rune",
            "in": "path",
            "description": "Rune ID or name (spaced or unspaced)",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "BITCOIN•PIZZAS"
          },
          {
            "name": "mempool",
            "in": "query",
            "description": "Mempool-aware",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServeResponse_Vec_RuneUtxo"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/addresses/{address}/utxos": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "UTxOs by Address",
        "description": "Returns UTxOs controlled by the provided address, sorted by height.",
        "operationId": "addresses_utxos_by_address",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Bitcoin address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "tb1qphcdyah2e4vtpxn56hsz3p6kapg90pl4x525kc"
          },
          {
            "name": "mempool",
            "in": "query",
            "description": "Mempool-aware",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServeResponse_Vec_AddressUtxo"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/runes/info": {
      "post": {
        "tags": [
          "Runes (Metaprotocol)"
        ],
        "summary": "Info by Rune (Batch)",
        "description": "Given a list of rune identifiers (name or id), returns a map of identifiers to rune info and a list of identifiers for which information could not be found.",
        "operationId": "runes_rune_info",
        "parameters": [
          {
            "name": "mempool",
            "in": "query",
            "description": "Mempool-aware",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServeResponse_RuneInfoBatch"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChainTip": {
        "type": "object",
        "required": [
          "block_hash",
          "block_height"
        ],
        "properties": {
          "block_hash": {
            "type": "string",
            "description": "The hash of the block"
          },
          "block_height": {
            "type": "integer",
            "format": "int64",
            "description": "The height of the block in the blockchain",
            "minimum": 0
          }
        }
      },
      "EstimatedBlock": {
        "type": "object",
        "required": [
          "block_height"
        ],
        "properties": {
          "block_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "IndexerInfo": {
        "type": "object",
        "required": [
          "chain_tip",
          "estimated_blocks"
        ],
        "properties": {
          "chain_tip": {
            "$ref": "#/components/schemas/ChainTip",
            "description": "Most recent *mined* block in the indexed mainchain, any estimated blocks will be descendants of this block"
          },
          "estimated_blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EstimatedBlock"
            },
            "description": "Information about any estimated blocks from the mempool that were indexed in addition to the mainchain"
          },
          "mempool_timestamp": {
            "type": [
              "string",
              "null"
            ],
            "description": "Timestamp of the indexed mempool snapshot, if any estimated blocks from the mempool have been indexed"
          }
        }
      },
...
      "RuneAndAmount": {
        "type": "object",
        "required": [
          "id",
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        }
      },
...
      "RuneInfo": {
        "type": "object",
        "required": [
          "id",
          "name",
          "spaced_name",
          "divisibility",
          "etching_tx",
          "etching_height",
          "premine"
        ],
        "properties": {
          "divisibility": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "etching_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "etching_tx": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "premine": {
            "type": "string"
          },
          "spaced_name": {
            "type": "string"
          },
          "symbol": {
            "type": [
              "string",
              "null"
            ]
          },
          "terms": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RuneTerms"
              }
            ]
          }
        }
      },
...
      "RuneTerms": {
        "type": "object",
        "properties": {
          "amount": {
            "type": [
              "string",
              "null"
            ]
          },
          "cap": {
            "type": [
              "string",
              "null"
            ]
          },
          "end_height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "start_height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "ServeResponse_RuneInfoBatch": {
        "type": "object",
        "required": [
          "data",
          "indexer_info"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "found",
              "missing"
            ],
            "properties": {
              "found": {
                "type": "object",
                "additionalProperties": {
                  "$ref": "#/components/schemas/RuneInfo"
                },
                "propertyNames": {
                  "type": "string"
                }
              },
              "missing": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "indexer_info": {
            "$ref": "#/components/schemas/IndexerInfo"
          }
        }
      },
      "ServeResponse_Vec_AddressUtxo": {
        "type": "object",
        "required": [
          "data",
          "indexer_info"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "tx_hash",
                "output_index",
                "height",
                "satoshis"
              ],
              "properties": {
                "height": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "output_index": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                },
                "satoshis": {
                  "type": "string"
                },
                "tx_hash": {
                  "type": "string"
                }
              }
            }
          },
          "indexer_info": {
            "$ref": "#/components/schemas/IndexerInfo"
          }
        }
      },
...
      "ServeResponse_Vec_RuneUtxo": {
        "type": "object",
        "required": [
          "data",
          "indexer_info"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "tx_hash",
                "output_index",
                "height",
                "satoshis",
                "runes"
              ],
              "properties": {
                "height": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "output_index": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                },
                "runes": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuneAndAmount"
                  }
                },
                "satoshis": {
                  "type": "string"
                },
                "tx_hash": {
                  "type": "string"
                }
              }
            }
          },
          "indexer_info": {
            "$ref": "#/components/schemas/IndexerInfo"
          }
        }
      }
    }
  }
}